Refactored the stars component in a new folder
