+++ src/components/stars/stars/EmptyStar.js
