+++ src/components/stars/stars/HalfStar.js
